Extract TransactionCard to new composable function
